# fitness-cms/web-cms/public/swagger.yaml
openapi: 3.0.3
info:
  title: FitToday CMS API
  description: |
    API for the FitToday fitness CMS platform. Connects personal trainers with students
    through programs, workouts, and real-time communication.

    ## Authentication
    All protected endpoints require a Firebase Auth token in the Authorization header:
    ```
    Authorization: Bearer <firebase-id-token>
    ```

    ## Roles
    - **trainer** - Personal trainers who create programs and manage students
    - **student** - Students who follow programs and track progress
    - **admin** - Platform administrators

    ## Plans
    | Plan    | Price (BRL) | Commission | Max Programs | Max Students |
    |---------|-------------|------------|--------------|--------------|
    | Starter | Free        | 10%        | 3            | 5            |
    | Pro     | R$97/mo     | 5%         | Unlimited    | Unlimited    |
    | Elite   | R$197/mo    | 0%         | Unlimited    | Unlimited    |
  version: 1.0.0
  contact:
    name: FitToday Engineering
servers:
  - url: https://fittoday.vercel.app
    description: Production
  - url: http://localhost:3001
    description: Local development
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: Firebase ID Token
      description: Firebase Authentication ID token
  schemas:
...
    ConnectionActionRequest:
      type: object
      required:
        - action
      properties:
        action:
          type: string
          enum: [accept, reject, cancel]
        reason:
          type: string
          maxLength: 500
          description: Optional reason (for cancel action)

    # ---- Reviews ----
...
    AdminTrainerListItem:
      type: object
      properties:
        uid:
          type: string
        email:
          type: string
        displayName:
          type: string
        photoURL:
          type: string
        status:
          type: string
          enum: [pending, active, suspended, rejected]
        createdAt:
          type: string
          format: date-time
        statusUpdatedAt:
          type: string
          format: date-time
        profile:
          type: object
          properties:
            bio:
              type: string
            specialties:
              type: array
              items:
                type: string
            experience:
              type: integer
        store:
          type: object
          properties:
            totalStudents:
              type: integer
            totalSales:
              type: integer
    AdminTrainerListResponse:
      type: object
      properties:
        trainers:
          type: array
          items:
            $ref: "#/components/schemas/AdminTrainerListItem"
        total:
          type: integer
        pending:
          type: integer
        active:
          type: integer
        suspended:
          type: integer
...
paths:
...
  /api/connections/{id}:
    patch:
      tags: [Connections]
      summary: Respond to a connection request
      description: |
        - **accept/reject**: Only the trainer who owns the connection can respond. Connection must be pending.
        - **cancel**: Either the trainer or student can cancel. If active, cleans up subscription, chat, and student link.
        Sends notifications to the other party.
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ConnectionActionRequest"
      responses:
        "200":
          description: Action completed
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: string
                  status:
                    type: string
                  subscriptionId:
                    type: string
                    description: Only on accept
                  chatRoomId:
                    type: string
                    description: Only on accept
                  cancelledBy:
                    type: string
                    description: Only on cancel
                  reason:
                    type: string
                    description: Only on cancel
        "400":
          description: Invalid action
        "401":
          description: Unauthorized
        "403":
          description: Forbidden
        "404":
          description: Connection not found
        "409":
          description: Already responded or already cancelled

  # ================================================================
  # PROGRAMS
  # ================================================================
...
  /api/chats:
    get:
      tags: [Chats]
      summary: Check if chat room exists
      description: Returns the chat room for a trainer-student pair. Only participants can query.
      security:
        - BearerAuth: []
      parameters:
        - name: trainerId
          in: query
          required: true
          schema:
            type: string
        - name: studentId
          in: query
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Chat room status
          content:
            application/json:
              schema:
                type: object
                properties:
                  exists:
                    type: boolean
                  chatRoom:
                    nullable: true
                    type: object
                    description: Chat room object or null if not found
                    properties:
                      id:
                        type: string
                      trainerId:
                        type: string
                      studentId:
                        type: string
                      isActive:
                        type: boolean
                      lastMessage:
                        type: string
                      createdAt:
                        type: string
                        format: date-time
        "400":
          description: Missing trainerId or studentId
        "401":
          description: Unauthorized
        "403":
          description: Not a participant

  # ================================================================
  # NOTIFICATIONS
  # ================================================================
...
  /api/admin/trainers:
    get:
      tags: [Admin]
      summary: List all trainers with filters
      description: Admin-only. Returns all trainers with status counts.
      security:
        - BearerAuth: []
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [pending, active, suspended, rejected, all]
        - name: search
          in: query
          schema:
            type: string
          description: Search by name or email
        - name: limit
          in: query
          schema:
            type: integer
            default: 50
      responses:
        "200":
          description: Trainer list with counts
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminTrainerListResponse"
        "401":
          description: Unauthorized (admin only)
...